detection-fissures/requirements.txt
fastapi==0.75.0
uvicorn==0.18.3

ultralytics==8.1.0
opencv-python-headless==4.8.1.78
pillow==9.5.0
numpy==1.24.3

torch==2.1.2+cpu
torchvision==0.16.2+cpu

python-multipart==0.0.9

--extra-index-url https://download.pytorch.org/whl/cpu
